Auto scale y-axis fixed

+++ app/src/main/java/com/jeeyo/sagar/jeeyodill/StreamplotJNIWrapper.java
@@ -13,5 +13,5 @@ public class StreamplotJNIWrapper {
 
     public static native void on_draw_frame();
 
-    public static native void add(int data);
+    public static native void add(float data);
 }
